requirements.txt
attrs==21.4.0
Authlib==1.0.1
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
idna==3.3
isodate==0.6.1
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
numpy==1.23.0
panda==0.3.1
pandas==1.4.3
platformdirs==2.5.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
requests-file==1.5.1
requests-toolbelt==0.9.1
simple-salesforce==1.11.6
six==1.16.0
urllib3==1.26.9
zeep==4.1.0
